The queue structure is modified to have separate queue for each capture object and scan manager queue to merge all
Now there is separate queue for each cpature object so that they dont block. And they all send data to the MC queue of scan manager

The scan manager struct is now opaque. internally now memory is provided to user from static array
Since the struct is opaque, the create function return the interface member which is public to the user.